backend/requirements.txt
# Flask and Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Machine Learning & Computer Vision
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0

# Reinforcement Learning (Future)
stable-baselines3==2.1.0

# Image Processing
Pillow==10.0.1

# Video Processing
ffmpeg-python==0.2.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pandas==2.0.3

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production
gunicorn==21.2.0
redis==4.6.0
celery==5.3.1 